Suggesting the POST CURL changes to support login
Suggesting the POST CURL changes to support login via credentials. In the existing code, it gets the error from WF because the data that we are passing from Login is using GET which is not WF wants. So by having this piece of code will solve that error and don't break the existing code as well.